+++ gateway/singleton.go
@@ -2,9 +2,13 @@ package gateway
 
 import (
 	"context"
+	"github.com/viant/afs"
+	"github.com/viant/afs/file"
+	"github.com/viant/afs/url"
 	"github.com/viant/datly/config"
 	"github.com/viant/gmetric"
 	"net/http"
+	"os"
 	"sync"
 )
 
@@ -13,10 +17,11 @@ var once sync.Once
 
 func Singleton(configURL string, statusHandler http.Handler, authorizer Authorizer, registry *config.Registry, metric *gmetric.Service) (*Service, error) {
 	var err error
+	fs := NewFs(configURL)
 	once.Do(func() {
 		ctx := context.Background()
 		var config *Config
-		if config, err = NewConfigFromURL(ctx, configURL); err != nil {
+		if config, err = NewConfigFromURL(ctx, fs, configURL); err != nil {
 			return
 		}
 		service, err = New(ctx, config, statusHandler, authorizer, registry, metric)
@@ -26,6 +31,14 @@ func Singleton(configURL string, statusHandler http.Handler, authorizer Authoriz
 	}
 
 	return service, err
+}
+
+func NewFs(configURL string) afs.Service {
+	if os.Getenv("DATLY_FS") == "cfs" {
+		ParentURL, _ := url.Split(configURL, file.Scheme)
+		return NewCacheFs(ParentURL)
+	}
+	return afs.New()
 }
 
 func SingletonWithConfig(config *Config, statusHandler http.Handler, authorizer Authorizer, registry *config.Registry, metric *gmetric.Service) (*Service, error) {